backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
gunicorn==21.2.0
PyJWT==2.8.0
scikit-learn==1.4.2
